+++ ThinkingHome.Core.Infrastructure/HomeApplication.cs
@@ -1,6 +1,8 @@
 ﻿using System;
+using System.Globalization;
 using System.Linq;
 using System.Reflection;
+using System.Threading;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
@@ -28,6 +30,8 @@ namespace ThinkingHome.Core.Infrastructure
 
             logger = loggerFactory.CreateLogger<HomeApplication>();
             context = services.GetRequiredService<IServiceContext>();
+
+            InitLanguage(config);
 
             try
             {
@@ -83,15 +87,30 @@ namespace ThinkingHome.Core.Infrastructure
 
         #region private
 
-        private IServiceProvider ConfigureServices(HomeConfiguration config)
+        private void InitLanguage(HomeConfiguration config)
+        {
+            var culture = config.GetCulture();
+
+            logger.LogInformation($"init culture: {culture}");
+
+            Thread.CurrentThread.CurrentCulture =
+                Thread.CurrentThread.CurrentUICulture =
+                    CultureInfo.DefaultThreadCurrentCulture =
+                        CultureInfo.DefaultThreadCurrentUICulture = culture;
+
+        }
+
+        private static IServiceProvider ConfigureServices(HomeConfiguration config)
         {
             var asms = config.GetDependencies().ToArray();
 
             var serviceCollection = new ServiceCollection();
 
+            serviceCollection.AddOptions();
             serviceCollection.AddSingleton<ILoggerFactory, LoggerFactory>();
             serviceCollection.AddSingleton<IServiceContext, ServiceContext>();
             serviceCollection.AddSingleton<IConfigurationSection>(config.Configuration.GetSection("plugins"));
+            serviceCollection.AddLocalization(opts => opts.ResourcesPath = "Lang");
 
             foreach (var asm in asms)
             {
@@ -101,7 +120,7 @@ namespace ThinkingHome.Core.Infrastructure
             return serviceCollection.BuildServiceProvider();
         }
 
-        private void AddAssemblyPlugins(ServiceCollection services, Assembly asm)
+        private static void AddAssemblyPlugins(ServiceCollection services, Assembly asm)
         {
             var baseType = typeof(PluginBase);