+++ ThinkingHome.Core.Plugins/IServiceContext.cs
@@ -1,13 +1,12 @@
-﻿using System;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
+using Microsoft.Extensions.Configuration;
+using Microsoft.Extensions.Localization;
 
 namespace ThinkingHome.Core.Plugins
 {
     public interface IServiceContext
     {
         IReadOnlyCollection<PluginBase> GetAllPlugins();
-
-        IReadOnlyCollection<T> GetAllPlugins<T>();
 
         T Require<T>() where T : PluginBase;
     }