+++ ThinkingHome.Plugins.Cron.WebUi/Resources/list.js
@@ -1,20 +1,22 @@
 var lib = require('lib');
+var lang = require('lang!static/cron/lang.json');
 
-var itemTemplate = 
+var itemTemplate =
     '<span class="th-cron-value">{{#isnull month}}*{{else}}{{month}}{{/isnull}}</span>:' +
     '<span class="th-cron-value">{{#isnull day}}*{{else}}{{day}}{{/isnull}}</span>:' +
     '<span class="th-cron-value">{{#isnull hour}}*{{else}}{{hour}}{{/isnull}}</span>:' +
     '<span class="th-cron-value">{{#isnull minute}}*{{else}}{{pad minute 2}}{{/isnull}}</span>' +
-    '<a href="#" class="js-task-edit">{{name}}</a> {{#unless enabled}}disabled{{/unless}}';
+    '<a href="#" class="js-task-edit">{{name}}</a> {{#unless enabled}}{{lang \'disabled\'}}{{/unless}}';
 
-var layoutTemplate = '<h1>Cron tasks</h1>' +
-    '<p><a href="#" class="btn btn-secondary js-task-create">Create</a></p>' +
+var layoutTemplate = '<h1>{{lang \'Cron tasks\'}}</h1>' +
+    '<p><a href="#" class="btn btn-secondary js-task-create">{{lang \'Create\'}}</a></p>' +
     '<div class="js-task-list"></div>';
 
 //#region views
 
 var ItemView = lib.marionette.View.extend({
     template: lib.handlebars.compile(itemTemplate),
+    templateContext: { lang: lang },
     className: 'th-list-item',
     tagName: 'li',
     triggers: {
@@ -30,6 +32,7 @@ var ListView = lib.marionette.CollectionView.extend({
 
 var LayoutView = lib.marionette.View.extend({
     template: lib.handlebars.compile(layoutTemplate),
+    templateContext: { lang: lang },
     regions: {
         list: '.js-task-list'
     },
@@ -44,7 +47,7 @@ var Section = lib.common.AppSection.extend({
     start: function() {
         this.view = new LayoutView();
         this.listenTo(this.view, 'task:create', this.bind('createTask'));
-        
+
         this.application.setContentView(this.view);
 
         return lib.ajax