+++ ThinkingHome.Plugins.WebServer/Attributes/HttpLocalizationResourceAttribute.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using ThinkingHome.Plugins.WebServer.Attributes.Base;
+
+namespace ThinkingHome.Plugins.WebServer.Attributes
+{
+    [AttributeUsage(AttributeTargets.Class)]
+    public class HttpLocalizationResourceAttribute : HttpResourceAttribute
+    {
+        public HttpLocalizationResourceAttribute(string url)
+            : base(url, "application/json;charset=utf-8", false)
+        {
+        }
+    }
+}