+++ ThinkingHome.Plugins.WebUi/Resources/Application/loaders/system-lang.js
@@ -1,0 +1,6 @@
+var lib = require('lib');
+
+exports.instantiate = function (load) {
+    var data = JSON.parse(load.source);
+    return new lib.common.StringLocalizer(data);
+};