+++ ThinkingHome.Console/CHANGES.md
@@ -70,6 +70,11 @@ WEB SERVER
   - `lib.ajax.getJSON(url, [data])`
   - `lib.ajax.postJSON(url, [data])`
   - обработка асинхронных исключений в метода start (нужно вернуть промис из него)
+  - локализация: вместо статических классов свойство `StringLocalizer` у плагина
+  - `[HttpLocalizationResource("/static/tmp/lang.json")]` - на клиент приходят переводы текущего плагина в json
+  - клиентская локализация: было `templateHelpers: { lang: lang }`, стало `templateContext: { lang: lang }`
+  - локализация: название раздела web UI берется из ресурсов текущего плагина
+
 - mail plugin: `sendMail` + `host.api.sendMail`, `host.api.sendMailWithAttachment`
 - MQTT плагин:
   - можно подписаться на несколько топиков